+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Lightstreamer Changelog
 
-### 0.2 — Unreleased
+### 0.2 — July 23, 2016
 
+- Added complete test suite
 - Added `Lightstreamer::Session#subscribed?` and `Lightstreamer::Session#disconnect`
 - Fixed `Lightstreamer::ProtocolError#code` not being set
 - Renamed the command-line client's `--address` option to `--server-url`